Pretend we always have >= 2 players in map generation (fixes #30)

+++ environment/core/hlt.hpp
@@ -194,7 +194,7 @@ namespace hlt {
         Map(const Map& other_map);
         Map(unsigned short width,
             unsigned short height,
-            uint8_t player_count,
+            unsigned int player_count,
             unsigned int seed);
 
         auto get_ship(PlayerId player, EntityIndex entity) -> Ship&;
